slate cleanup
